Update capture script: drawing caption bug
captube: fix a drawing caption bug
If the lastCut indicate a last charactor of the string, then it will be
IndexError by 'nextCut += 1'. And also if the lastCut indicate a first
charactor of the string, then it will be IndexError by 'nextCut -= 1'.
The bugs were occured well especially for the no space language like
Chinese.

Signed-off-by: Ji-Hun Kim <jihuun.k@gmail.com>